scrypts/posgreSQL-exporter/requirements.txt
pika==1.3.1
psycopg2-binary==2.9.5
